backend/requirements.txt
# Core
fastapi[standard]
python-dotenv
pydantic-settings
gunicorn

# AI Providers
langchain
langchain-openai
langchain-google-genai
google-genai
langchain-anthropic
anthropic
langchain-xai
langchain-deepseek
requests

# Database
sqlalchemy[asyncio]
psycopg2-binary
alembic
asyncpg
aiosqlite
